+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/XR/XRSystem.cs
@@ -104,9 +104,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             {
                 display = null;
             }
-#endif
 
             return display != null;
+#endif
+
+            return false;
         }
 
         void CreateLayoutLegacyStereo(Camera camera)