+++ com.unity.shadergraph/Editor/Data/Nodes/MasterNode.cs
@@ -6,7 +6,6 @@ using UnityEditor.Graphing.Util;
 using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.UIElements;
-using UnityEditor.ShaderGraph.Drawing.Inspector;
 
 namespace UnityEditor.ShaderGraph
 {